requirements.txt
Django<1.12
psycopg2
ipython
Pillow
django-ckeditor